Reorganize for legacy tests
